Global and Post-specific settings
Global defaults can be set in the settings section.
Post-specific settings can be adjusted in the meta box when editing a
post.
Added “Posted via” setting, to set a custom “via” value to be displayed
when posting to Diaspora*.
Modified logic if a password has already been entered.
Issues #14 and #17.